code cleanup

+++ wcd/src/c3po/Text.c
@@ -27,7 +27,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include "nameset.h"
 #include "intset.h"
 
-expfun text textNewSize(int size)
+expfun text textNewSize(size_t size)
 {
    text t = NULL;
